requirements.txt
Django==4.0.3
psycopg2-binary==2.9.3
requests==2.27.1
django-filter
crispy-bootstrap5
djangorestframework
flake8
